put validation and stepper form in separate files
- add validation resource to readme
- move stepper form logic to one file
- move validation logic to one file
- update import and add type check to script.js
- remove unused classes from signup forms
