+++ templates/base.html
@@ -70,7 +70,7 @@
                                 <summary>
                                     Account
                                 </summary>
-                                <ul class="p-2 bg-base-100">
+                                <ul class="p-2 bg-base-100 z-50">
                                     <li>
                                         <a href="{% url 'user_profile' slug=user.slug %}"
                                             aria-label="Navigates to your profile page">My profile</a>
